// sdk/src/types.ts
export type VerificationMethod = {
  /**
   * A multikey that specificies a public key for the verification method.
   * @see https://www.w3.org/TR/vc-data-integrity/#multikey
   */
  multikey: Buffer;
  /**  The verification relationships for this verification method. */
  verificationRelationshipFlags: VerificationRelationshipFlags;
};

export type Did = {
  /**
   * The verification methods for the DID.
   * @see https://www.w3.org/TR/did-core/#verification-methods
   */
  verificationMethods: VerificationMethod[];
  /** The p2tr key corresponding to the DID UTXO that can be used to update or deactivate the DID. */
  controllerKey?: Buffer;
  /** Additional metadata pertaining to the DID such as services. */
  metadata?: Record<string, object>;
  isDeactivated?: boolean;
};
// ...
export type VerificationMethodDeletion = {
  i: number;
};
export type VerificationMethodUpdate = {
  i: number;
  k?: string;
  vr?: VerificationRelationshipFlags;
};
export type VerificationMethodAppend = {
  k: string;
  vr: VerificationRelationshipFlags;
};
export type VerificationMethodOperation =
  | VerificationMethodDeletion
  | VerificationMethodUpdate
  | VerificationMethodAppend;

/**
 * @see https://microstrategy.github.io/did-btc-spec/#update
 */
export type DidUpdate = {
  /** Changes to the verification methods for the DID. */
  vm?: VerificationMethodOperation[];
  /** Metadata records to update in the DID document */
  u?: Record<string, object>;
  /** Metadata to delete from the DID document. */
  d?: Record<string, object>;
  /** Metadata to add to the DID document. */
  a?: Record<string, object>;
};